♻️ Refactor admin interface into modular structure
null
