+++ cmd/experiments/js-web-server/internal/web/static/admin/globalstate.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>GlobalState Inspector - Admin Console</title>
+    <link rel="stylesheet" href="/admin/globalstate.css">
+</head>
+<body>
+    <div class="header">
+        <h1>GlobalState Inspector</h1>
+        <div class="nav-links">
+            <a href="/admin/logs">Request Logs</a>
+            <a href="/admin/scripts">Scripts</a>
+            <a href="/">Playground</a>
+        </div>
+    </div>
+    
+    <div class="controls">
+        <button onclick="refreshGlobalState()">Refresh</button>
+        <button onclick="saveGlobalState()" class="success">Save Changes</button>
+        <button onclick="resetGlobalState()" class="danger">Reset to {}</button>
+        <div class="auto-refresh">
+            <input type="checkbox" id="autoRefresh" onchange="toggleAutoRefresh()">
+            <label for="autoRefresh">Auto-refresh (5s)</label>
+        </div>
+    </div>
+    
+    <div class="main-content">
+        <div class="editor-container">
+            <div class="editor-header">
+                GlobalState JSON Editor
+            </div>
+            <div class="editor">
+                <textarea id="globalStateEditor" placeholder="Loading globalState..."></textarea>
+            </div>
+            <div class="status-bar">
+                <div>
+                    <span class="validation-status" id="validationStatus">Valid JSON</span>
+                    <span id="errorMessage"></span>
+                </div>
+                <div>
+                    Last updated: <span id="lastUpdated">Never</span>
+                </div>
+            </div>
+        </div>
+        
+        <div class="help-panel">
+            <div class="help-header">
+                Help & Usage
+            </div>
+            <div class="help-content">
+                <p>The <code>globalState</code> object persists data across JavaScript executions in the playground. Use it to:</p>
+                <ul>
+                    <li>Store configuration values</li>
+                    <li>Maintain counters or statistics</li>
+                    <li>Share data between different endpoints</li>
+                    <li>Cache computed results</li>
+                </ul>
+                <p><strong>Example usage in JavaScript:</strong></p>
+                <code>
+                    globalState.counter = (globalState.counter || 0) + 1;<br>
+                    globalState.users = globalState.users || [];<br>
+                    globalState.config = { theme: 'dark', version: '1.0' };
+                </code>
+            </div>
+        </div>
+    </div>
+    
+    <div class="notification" id="notification"></div>
+
+    <script src="/admin/globalstate.js"></script>
+</body>
+</html>
